requirements.txt
pandas
numpy
seaborn
matplotlib
Pillow
tensorflow
keras
imageio
opencv-python
pydub
scipy
librosa
xgboost
lightgbm
catboost
scikit-learn
ipython
git+https://github.com/openai/whisper.git